[UserManagementBundle] use form to delete user

+++ src/Kunstmaan/UserManagementBundle/Controller/UsersController.php
@@ -3,21 +3,24 @@
 namespace Kunstmaan\UserManagementBundle\Controller;
 
 use Doctrine\ORM\EntityManager;
+use Exception;
 use FOS\UserBundle\Event\UserEvent;
 use FOS\UserBundle\Model\UserInterface;
 use Kunstmaan\AdminBundle\Controller\BaseSettingsController;
 use Kunstmaan\AdminBundle\Entity\BaseUser;
+use Kunstmaan\AdminBundle\Entity\User;
 use Kunstmaan\AdminBundle\Event\AdaptSimpleFormEvent;
 use Kunstmaan\AdminBundle\Event\Events;
 use Kunstmaan\AdminBundle\FlashMessages\FlashTypes;
 use Kunstmaan\AdminBundle\Form\RoleDependentUserFormInterface;
 use Kunstmaan\AdminListBundle\AdminList\AdminList;
 use Kunstmaan\UserManagementBundle\Event\UserEvents;
-use Symfony\Component\Routing\Annotation\Route;
+use Kunstmaan\UserManagementBundle\Form\UserDeleteForm;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
@@ -30,8 +33,6 @@ class UsersController extends BaseSettingsController
      *
      * @Route("/", name="KunstmaanUserManagementBundle_settings_users")
      * @Template("@KunstmaanAdminList/Default/list.html.twig")
-     *
-     * @param \Symfony\Component\HttpFoundation\Request $request
      *
      * @return array
      */
@@ -53,9 +54,9 @@ class UsersController extends BaseSettingsController
         $adminList = $this->container->get('kunstmaan_adminlist.factory')->createList($configurator);
         $adminList->bindRequest($request);
 
-        return array(
+        return [
             'adminlist' => $adminList,
-        );
+        ];
     }
 
     /**
@@ -76,8 +77,6 @@ class UsersController extends BaseSettingsController
      * @Route("/add", name="KunstmaanUserManagementBundle_settings_users_add", methods={"GET", "POST"})
      * @Template("@KunstmaanUserManagement/Users/add.html.twig")
      *
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     *
      * @return array
      */
     public function addAction(Request $request)
@@ -86,7 +85,7 @@ class UsersController extends BaseSettingsController
 
         $user = $this->getUserClassInstance();
 
-        $options = array('password_required' => true, 'langs' => $this->container->getParameter('kunstmaan_admin.admin_locales'), 'validation_groups' => array('Registration'), 'data_class' => \get_class($user));
+        $options = ['password_required' => true, 'langs' => $this->container->getParameter('kunstmaan_admin.admin_locales'), 'validation_groups' => ['Registration'], 'data_class' => \get_class($user)];
         $formTypeClassName = $user->getFormTypeClass();
         $formType = new $formTypeClassName();
 
@@ -120,9 +119,9 @@ class UsersController extends BaseSettingsController
             }
         }
 
-        return array(
+        return [
             'form' => $form->createView(),
-        );
+        ];
     }
 
     /**
@@ -133,9 +132,9 @@ class UsersController extends BaseSettingsController
      * @Route("/{id}/edit", requirements={"id" = "\d+"}, name="KunstmaanUserManagementBundle_settings_users_edit", methods={"GET", "POST"})
      * @Template("@KunstmaanUserManagement/Users/edit.html.twig")
      *
-     * @throws AccessDeniedException
-     *
      * @return array
+     *
+     * @throws AccessDeniedException
      */
     public function editAction(Request $request, $id)
     {
@@ -159,7 +158,7 @@ class UsersController extends BaseSettingsController
         $userEvent = new UserEvent($user, $request);
         $this->container->get('event_dispatcher')->dispatch(UserEvents::USER_EDIT_INITIALIZE, $userEvent);
 
-        $options = array('password_required' => false, 'langs' => $this->container->getParameter('kunstmaan_admin.admin_locales'), 'data_class' => \get_class($user));
+        $options = ['password_required' => false, 'langs' => $this->container->getParameter('kunstmaan_admin.admin_locales'), 'data_class' => \get_class($user)];
         $formFqn = $user->getFormTypeClass();
         $formType = new $formFqn();
 
@@ -197,57 +196,55 @@ class UsersController extends BaseSettingsController
                 return new RedirectResponse(
                     $this->generateUrl(
                         'KunstmaanUserManagementBundle_settings_users_edit',
-                        array('id' => $id)
+                        ['id' => $id]
                     )
                 );
             }
         }
 
-        $params = array(
+        $params = [
             'form' => $form->createView(),
             'user' => $user,
-        );
+        ];
 
         if ($tabPane) {
-            $params = array_merge($params, array('tabPane' => $tabPane));
+            $params = array_merge($params, ['tabPane' => $tabPane]);
         }
 
         return $params;
     }
 
     /**
-     * Delete a user
+     * @Route("/delete", requirements={"id" = "\d+"}, name="KunstmaanUserManagementBundle_settings_users_delete", methods={"POST"})
      *
-     * @param Request $request
-     * @param int     $id
-     *
-     * @Route("/{id}/delete", requirements={"id" = "\d+"}, name="KunstmaanUserManagementBundle_settings_users_delete", methods={"POST"})
-     *
-     * @throws AccessDeniedException
-     *
-     * @return array
+     * @throws Exception
      */
-    public function deleteAction(Request $request, $id)
+    public function deleteAction(Request $request)
     {
         $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
 
         /* @var EntityManager $em */
         $em = $this->getDoctrine()->getManager();
-        /* @var UserInterface $user */
-        $user = $em->getRepository($this->container->getParameter('fos_user.model.user.class'))->find($id);
-        if (!\is_null($user)) {
-            $userEvent = new UserEvent($user, $request);
-            $this->container->get('event_dispatcher')->dispatch(UserEvents::USER_DELETE_INITIALIZE, $userEvent);
+        $user = new User();
+        $form = $this->createForm(UserDeleteForm::class, $user);
+        $form->submit($request->request->get($form->getName()));
+        if ($form->isSubmitted() && $form->isValid()) {
+            /* @var UserInterface $user */
+            $user = $em->getRepository($this->container->getParameter('fos_user.model.user.class'))->find($user->getId());
+            if (!\is_null($user)) {
+                $userEvent = new UserEvent($user, $request);
+                $this->container->get('event_dispatcher')->dispatch(UserEvents::USER_DELETE_INITIALIZE, $userEvent);
 
-            $em->remove($user);
-            $em->flush();
+                $em->remove($user);
+                $em->flush();
 
-            $this->addFlash(
-                FlashTypes::SUCCESS,
-                $this->container->get('translator')->trans('kuma_user.users.delete.flash.success.%username%', [
-                    '%username%' => $user->getUsername(),
-                ])
-            );
+                $this->addFlash(
+                    FlashTypes::SUCCESS,
+                    $this->container->get('translator')->trans('kuma_user.users.delete.flash.success.%username%', [
+                        '%username%' => $user->getUsername(),
+                    ])
+                );
+            }
         }
 
         return new RedirectResponse($this->generateUrl('KunstmaanUserManagementBundle_settings_users'));
@@ -262,9 +259,9 @@ class UsersController extends BaseSettingsController
         return new RedirectResponse(
             $this->generateUrl(
                 'KunstmaanUserManagementBundle_settings_users_edit',
-                array(
+                [
                     'id' => $this->container->get('security.token_storage')->getToken()->getUser()->getId(),
-                )
+                ]
             )
         );
     }
